requirements.txt
Django
pillow
djangorestframework